+++ src/actions/mcubootTargetActions.ts
@@ -6,8 +6,13 @@
 
 /* eslint-disable import/no-cycle */
 
-import nrfdl from '@nordicsemiconductor/nrf-device-lib-js';
-import { Device, getDeviceLibContext, logger } from 'pc-nrfconnect-shared';
+import nrfdl, { SerialPort } from '@nordicsemiconductor/nrf-device-lib-js';
+import {
+    Device,
+    getDeviceLibContext,
+    logger,
+    usageData,
+} from 'pc-nrfconnect-shared';
 
 import {
     mcubootFirmwareValid,
@@ -28,38 +33,111 @@ import {
 } from '../reducers/targetReducer';
 import { RootState, TDispatch } from '../reducers/types';
 import { targetWarningRemove } from '../reducers/warningReducer';
-import { CommunicationType } from '../util/devices';
+import {
+    CommunicationType,
+    DeviceFamily,
+    McubootProductIds,
+    ModemProductIds,
+    VendorId,
+} from '../util/devices';
 import { updateTargetWritable } from './targetActions';
+import EventAction from './usageDataActions';
 
 export const first = <T>(items: T[]): T | undefined => items[0];
 export const last = <T>(items: T[]): T | undefined => items.slice(-1)[0];
 
-export const openDevice = (selectedDevice: Device) => (dispatch: TDispatch) => {
-    const { serialPorts } = selectedDevice;
-    dispatch(
-        targetTypeKnown({
-            targetType: CommunicationType.MCUBOOT,
-            isRecoverable: false,
-        })
-    );
-    dispatch(mcubootKnown(true));
-    dispatch(modemKnown(true));
-    dispatch(
-        mcubootPortKnown({
-            port: first(serialPorts)?.comName ?? undefined,
-            port2: last(serialPorts)?.comName ?? undefined,
-        })
-    );
-    dispatch(updateTargetWritable());
-    dispatch(loadingEnd());
-};
+/**
+ * Check whether the device is MCUboot device or not by providing vender Id and product Id
+ *
+ * @param {number} vid Vender Id
+ * @param {number} pid Product Id
+ * @returns {boolean} whether the device is MCUboot device
+ */
+export const isMcuboot = (vid?: number, pid?: number) =>
+    vid &&
+    pid &&
+    vid === VendorId.NORDIC_SEMICONDUCTOR &&
+    McubootProductIds.includes(pid);
+
+/**
+ * Check whether the device is MCUboot device with modem or not by providing vender Id and product Id
+ *
+ * @param {number} vid Vender Id
+ * @param {number} pid Product Id
+ * @returns {boolean} whether the device is MCUboot device with modem
+ */
+export const isMcubootModem = (vid?: number, pid?: number) =>
+    vid &&
+    pid &&
+    vid === VendorId.NORDIC_SEMICONDUCTOR &&
+    ModemProductIds.includes(pid);
+
+export const openDevice =
+    () => (dispatch: TDispatch, getState: () => RootState) => {
+        const { device: inputDevice } = getState().app.target;
+        const device = inputDevice as Device;
+        const { serialPorts } = device;
+        const serialport = serialPorts[0];
+        const { vendorId, productId } = serialport as SerialPort;
+        const vid = vendorId ? parseInt(vendorId.toString(), 16) : undefined;
+        const pid = productId ? parseInt(productId.toString(), 16) : undefined;
+
+        dispatch(
+            targetTypeKnown({
+                targetType: CommunicationType.MCUBOOT,
+                isRecoverable: false,
+            })
+        );
+        dispatch(mcubootKnown(true));
+        if (isMcubootModem(vid, pid)) {
+            // Only Thingy91 matches the condition
+            // Update when a new Nordic USB device has both mcuboot and modem
+            dispatch(modemKnown(true));
+            usageData.sendUsageData(
+                EventAction.OPEN_DEVICE_FAMILY,
+                DeviceFamily.NRF91
+            );
+            usageData.sendUsageData(
+                EventAction.OPEN_DEVICE_VERSION,
+                'Thingy91'
+            );
+            usageData.sendUsageData(
+                EventAction.OPEN_DEVICE_BOARD_VERSION,
+                'PCA20035'
+            );
+        } else {
+            // Only Thingy53 matches the condition
+            // Update when a new Nordic USB device has mcuboot without modem
+            dispatch(modemKnown(false));
+            usageData.sendUsageData(
+                EventAction.OPEN_DEVICE_FAMILY,
+                DeviceFamily.NRF53
+            );
+            usageData.sendUsageData(
+                EventAction.OPEN_DEVICE_VERSION,
+                'Thingy53'
+            );
+            usageData.sendUsageData(
+                EventAction.OPEN_DEVICE_BOARD_VERSION,
+                'PCA20053'
+            );
+        }
+        dispatch(
+            mcubootPortKnown({
+                port: first(serialPorts)?.comName ?? undefined,
+                port2: last(serialPorts)?.comName ?? undefined,
+            })
+        );
+        dispatch(updateTargetWritable());
+        dispatch(loadingEnd());
+    };
 
 export const toggleMcuboot =
     () => (dispatch: TDispatch, getState: () => RootState) => {
         const { port } = getState().app.target;
-        const { isMcuboot } = getState().app.mcuboot;
+        const { isMcuboot: isMcubootTarget } = getState().app.mcuboot;
 
-        if (isMcuboot) {
+        if (isMcubootTarget) {
             dispatch(mcubootKnown(false));
             dispatch(mcubootPortKnown({}));
         } else {
@@ -82,15 +160,15 @@ export const canWrite =
 
         // Check if mcu firmware is detected.
         // If not, then return.
-        const { mcubootFilePath } = getState().app.file;
-        if (!mcubootFilePath) {
+        const { mcubootFilePath, zipFilePath } = getState().app.file;
+        if (!mcubootFilePath && !zipFilePath) {
             return;
         }
 
         // Check if target is MCU target.
         // If not, then return.
-        const { isMcuboot } = getState().app.mcuboot;
-        if (!isMcuboot) {
+        const { isMcuboot: isMcubootTarget } = getState().app.mcuboot;
+        if (!isMcubootTarget) {
             return;
         }
 
@@ -109,10 +187,14 @@ export const performUpdate =
         new Promise<void>(resolve => {
             const { device: inputDevice } = getState().app.target;
             const device = inputDevice as Device;
-            const { mcubootFilePath } = getState().app.file;
+            const { mcubootFilePath, zipFilePath } = getState().app.file;
+            const dfuFilePath = mcubootFilePath || zipFilePath;
+            const firmwareFormat = mcubootFilePath
+                ? 'NRFDL_FW_MCUBOOT'
+                : 'NRFDL_FW_MCUBOOT_MULTI_IMAGE';
 
             logger.info(
-                `Writing ${mcubootFilePath} to device ${device.serialNumber}`
+                `Writing ${dfuFilePath} to device ${device.serialNumber}`
             );
 
             const errorCallback = (error: nrfdl.Error) => {
@@ -156,8 +238,8 @@ export const performUpdate =
                 getDeviceLibContext(),
                 device.id,
                 'NRFDL_FW_FILE',
-                'NRFDL_FW_MCUBOOT',
-                mcubootFilePath as string,
+                firmwareFormat,
+                dfuFilePath as string,
                 completeCallback,
                 progressCallback
             );